requirements.txt
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==4.4.2
Flask==2.2.2
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
moviepy==1.0.3
numpy==1.23.2
oauthlib==3.2.0
outcome==1.2.0
Pillow==9.2.0
playwright==1.25.2
proglog==0.1.10
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
social-post-api==1.1.0
sortedcontainers==2.4.0
TikTokApi==5.2.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
Wand==0.6.10
webdriver-manager==3.8.3
websockets==10.1
Werkzeug==2.2.2
wsproto==1.2.0
